Introduce FileSystem interface for TypeScript SDK
Extract a FileSystem interface from the existing Filesystem class to enable
implementing different filesystem backends (HostFS, OverlayFS).
