+++ sdk/typescript/src/index_browser.ts
@@ -1,7 +1,7 @@
 import { AgentFSCore } from "./agentfs.js";
 import { DatabasePromise } from "@tursodatabase/database-common";
 import { KvStore } from "./kvstore.js";
-import { Filesystem } from "./filesystem.js";
+import { AgentFS as AgentFSImpl } from "./filesystem/index.js";
 import { ToolCalls } from "./toolcalls.js";
 import { Buffer } from "buffer";
 
@@ -9,7 +9,7 @@ export class AgentFS extends AgentFSCore {
     static async openWith(db: DatabasePromise): Promise<AgentFSCore> {
         const [kv, fs, tools] = await Promise.all([
             KvStore.fromDatabase(db),
-            Filesystem.fromDatabase(db, Buffer),
+            AgentFSImpl.fromDatabase(db, Buffer),
             ToolCalls.fromDatabase(db),
         ]);
         return new AgentFS(db, kv, fs, tools);
@@ -18,7 +18,7 @@ export class AgentFS extends AgentFSCore {
 
 export { AgentFSOptions } from './agentfs.js';
 export { KvStore } from './kvstore.js';
-export { Filesystem } from './filesystem.js';
-export type { Stats } from './filesystem.js';
+export { AgentFS as Filesystem } from './filesystem/index.js';
+export type { Stats, DirEntry, FilesystemStats, FileHandle, FileSystem } from './filesystem/index.js';
 export { ToolCalls } from './toolcalls.js';
 export type { ToolCall, ToolCallStats } from './toolcalls.js';
